Adding ability to add value to parent attribute
This change introduces '+' to the end of attribute names as sign
to add this value to parent attribute of the same name without '+'.

This adding is type specific, child attribute will always have the
same type as parent. Here is example with type string:
tag: value1
/child
 tag+: value2
Is the same as:
/child:
 tag: value1value2

Also there is more test coverage of fmf/base.py in this commit.

Documentation has been added for this feature to features.rst.

Warning: It is no longer possible to run tree.grow() on the same path.
         Doing so results in adding the same value again.