Extend the `filter()` documentation slightly
People tend to try using parentheses in their filter expressions
for obvious reasons. But it is not possible. I feel like this
should be mentioned in the documentation. In fact, with DNF you
never need parentheses to express any expression and I think this
fact also needs to be noted there.

Signed-off-by: Ondrej Moris <omoris@redhat.com>
Co-authored-by: Petr Šplíchal <psplicha@redhat.com>